Rename the gem from zawgyidetector to myanmar-tools
