+++ clients/ruby/lib/myanmar-tools/binary_markov.rb
