+++ clients/ruby/lib/myanmar-tools/resources/compatibility.tsv
