+++ clients/ruby/lib/myanmar-tools/resources/zawgyiUnicodeModel.dat
