+++ clients/ruby/lib/myanmar-tools/version.rb
@@ -1,0 +1,3 @@
+module MyanmarTools
+  VERSION = '0.1.0'
+end
