--- clients/ruby/lib/zawgyidetector/version.rb
@@ -1,3 +1,0 @@
-module Zawgyidetector
-  VERSION = "0.1.0"
-end
